+++ rinokeras/core/v1x/models/rmc.py
@@ -6,10 +6,10 @@ from tensorflow.keras import Model
 from tensorflow.keras.layers import RNN, Flatten, Reshape, Input, LSTMCell
 import numpy as np
 
-import rinokeras.python.v1x as rk
-from rinokeras.python.v1x.common.layers import WeightNormDense as Dense
-from rinokeras.python.v1x.common.attention import AttentionMap, ScaledDotProductSimilarity, AttentionQKVProjection
-from rinokeras.python.v1x.common.layers import PositionEmbedding, LearnedEmbedding, LayerDropout, \
+import rinokeras.core.v1x as rk
+from rinokeras.core.v1x.common.layers import WeightNormDense as Dense
+from rinokeras.core.v1x.common.attention import AttentionMap, ScaledDotProductSimilarity, AttentionQKVProjection
+from rinokeras.core.v1x.common.layers import PositionEmbedding, LearnedEmbedding, LayerDropout, \
     LayerNorm, DenseStack, Dropout
 
 from .transformer import TransformerEncoderBlock, TransformerMultiAttention, TransformerFeedForward