+++ rinokeras/core/v1x/models/transformer.py
@@ -8,13 +8,13 @@ from tensorflow.keras import Model
 from tensorflow.keras.layers import (BatchNormalization, Dropout, Embedding,
                                      Lambda, Conv1D)
 
-import rinokeras.python.v1x as rk
-from rinokeras.python.v1x.common.attention import MultiHeadAttention, SelfAttention
-from rinokeras.python.v1x.common.layers import WeightNormDense as Dense
-from rinokeras.python.v1x.common.layers import (DenseStack, EmbeddingTranspose,
+import rinokeras.core.v1x as rk
+from rinokeras.core.v1x.common.attention import MultiHeadAttention, SelfAttention
+from rinokeras.core.v1x.common.layers import WeightNormDense as Dense
+from rinokeras.core.v1x.common.layers import (DenseStack, EmbeddingTranspose,
                                      LayerDropout, LayerNorm,
                                      PositionEmbedding, Stack)
-from rinokeras.python.v1x.utils import get_shape
+from rinokeras.core.v1x.utils import get_shape
 
 DecoderResult = namedtuple('DecoderResult', [
                            'seqpos', 'inputs', 'cache', 'output_sequence', 'is_finished'])