+++ rinokeras/core/v1x/train/RinokerasGraph.py
@@ -6,7 +6,7 @@ import tensorflow as tf
 from tensorflow.python.client import timeline
 from tqdm import tqdm
 
-from rinokeras.python.v1x.utils import MetricsAccumulator
+from rinokeras.core.v1x.utils import MetricsAccumulator
 from .train_utils import Inputs