+++ rinokeras/python/v1x/__init__.py
@@ -1,6 +1,0 @@
-# import rinokeras.python.v1x.utils  # noqa: F401
-# import rinokeras.python.v1x.common  # noqa: F401
-# import rinokeras.models  # noqa: F401
-# import rinokeras.python.v1x.train  # noqa: F401
-
-# __all__ = ['common', 'models', 'rl', 'test', 'utils', 'train']