+++ rinokeras/testing/v1x/models/test_transformer.py
@@ -3,7 +3,7 @@ import numpy as np
 import tensorflow as tf
 import sys
 
-from rinokeras.python.v1x.models.transformer import Transformer
+from rinokeras.core.v1x.models.transformer import Transformer
 
 
 def test_transformer():